Move `registerPaths` and `setPathParser` to opt-in extensions
