+++ src/index.test.tsx
@@ -15,7 +15,6 @@ describe("next-overridable-hook", () => {
       events: expect.any(Object),
       async: expect.any(Boolean),
       _setCurrentUrl: expect.any(Function),
-      registerPaths: expect.any(Function),
       // Ensure the router has exactly these properties:
       asPath: "",
       basePath: "",
